fix: separation of components and performance optimization
