+++ Murugavel/sitback-furniture/package.json
@@ -13,7 +13,8 @@
     "fast-deep-equal": "^3.1.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-router-dom": "^7.5.3"
+    "react-router-dom": "^7.5.3",
+    "react-spinners": "^0.17.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.22.0",
