+++ Murugavel/sitback-furniture/src/components/loader/Loader.jsx
@@ -1,0 +1,11 @@
+import './loader.css'
+
+const Loader = ({ children }) => {
+    return (
+        <div className="loader">
+            {children}
+        </div>
+    )
+}
+
+export default Loader;
