+++ Murugavel/sitback-furniture/src/components/loader/loader.css
@@ -1,0 +1,4 @@
+.loader{
+    display: grid;
+    place-items: center;
+}
