+++ Murugavel/sitback-furniture/src/containers/card-list-container/card-list-container.css
@@ -7,3 +7,7 @@
     flex-wrap: wrap;
     margin: 2.8rem 2rem 1rem 3rem;
 }
+
+.card-list > .loader{
+    width: 23rem;
+}
