+++ Murugavel/sitback-furniture/src/containers/cart-container/CartContainer.jsx
@@ -1,16 +1,13 @@
-import CartItemCard from '../../components/cart-item-card/CartItemCard';
 import { useCartStateContext } from '../../context/CartContext';
 import './cart-container.css'
-import Button from '../../components/button/Button';
-import { useNavigate } from 'react-router-dom';
+import { lazy, Suspense } from 'react';
+import Loader from '../../components/loader/Loader';
+import { ClipLoader } from 'react-spinners';
+
+const CartDetailsContainer = lazy(() => import('../cart-details-container/CartDetailsContainer'));
 
 const CartContainer = () => {
     const cartData = useCartStateContext();
-    const navigate = useNavigate();
-    
-    const handleClick = () => {
-        navigate('/confirmOrder');
-    }
 
     if (cartData.cartItems.length === 0) {
         return null;
@@ -19,16 +16,9 @@ const CartContainer = () => {
     return (
         <div className="cart-container">
             <h2 className="cart-text">MY CART</h2>
-            <div className='cart-item-list'>
-                {cartData.cartItems.map((orderData, index) => <CartItemCard key={index} productData={orderData} />)}
-            </div>
-            <div className="amount-banner">
-                <div className="amount-details">
-                    <p className='amount-text'>TOTAL AMOUNT</p>
-                    <p><span className="rupee">₹ </span> {cartData.totalAmount}</p>
-                </div>
-                <Button handleClick={handleClick}>PLACE ORDER</Button>
-            </div>
+            <Suspense fallback={<Loader><ClipLoader size={100} color={'var(--loader)'} /></Loader>}>
+                <CartDetailsContainer cartData={cartData} />
+            </Suspense>
         </div>
     )
 }
