+++ Murugavel/sitback-furniture/src/containers/cart-details-container/CartDetailsContainer.jsx
@@ -1,0 +1,28 @@
+import { useNavigate } from "react-router-dom";
+import Button from "../../components/button/Button";
+import CartItemCard from "../../components/cart-item-card/CartItemCard";
+import './cart-details-container.css'
+
+const CartDetailsContainer = ({cartData}) => {
+    const navigate = useNavigate();
+    
+    const handleClick = () => {
+        navigate('/confirmOrder');
+    }
+    return (
+        <>
+            <div className='cart-item-list'>
+                {cartData.cartItems.map((orderData, index) => <CartItemCard key={index} productData={orderData} />)}
+            </div>
+            <div className="amount-banner">
+                <div className="amount-details">
+                    <p className='amount-text'>TOTAL AMOUNT</p>
+                    <p><span className="rupee">₹ </span> {cartData.totalAmount}</p>
+                </div>
+            <Button handleClick={handleClick}>PLACE ORDER</Button>
+            </div>
+        </>
+    )
+}
+
+export default CartDetailsContainer;
