+++ Murugavel/sitback-furniture/src/containers/cart-details-container/cart-details-container.css
@@ -1,0 +1,24 @@
+.cart-item-list{
+    display: flex;
+    flex-direction: column;
+    height: calc(100vh - 17.6rem);
+    overflow-y: scroll;
+    row-gap: 2rem;
+    margin: 1.7rem 2.35rem;
+    padding-right: 0.2rem;
+}
+
+.amount-banner{
+    height: 5.2rem;
+    background-color: var(--primary-color);
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 2.2rem;
+    box-shadow: 0px 2px 7px 0px var(--shadow);
+}
+
+.amount-text{
+    font-size: 0.75rem;
+    font-weight: 350;
+}
